feature: major refactor for better dev exp - functionality stays (except deprecated migrate which is removed)
